Working Replay POC using osmdroid with some bugs

+++ app/src/main/java/ch/epfl/sdp/replay/game_event/CatchEvent.kt
@@ -1,3 +1,3 @@
 package ch.epfl.sdp.replay.game_event
 
-class CatchEvent(timestamp: Long, val predatorID: Int, val preyID: Int) : GameEvent(timestamp)
+class CatchEvent(timestamp: Int, val predatorID: Int, val preyID: Int) : GameEvent(timestamp)
