+++ app/src/main/java/ch/epfl/sdp/replay/game_event/GameEvent.kt
@@ -2,4 +2,4 @@ package ch.epfl.sdp.replay.game_event
 
 import java.io.Serializable
 
-abstract class GameEvent(val timestamp: Long) : Serializable
+abstract class GameEvent(val timestamp: Int) : Serializable
