+++ app/src/main/java/ch/epfl/sdp/replay/game_event/LocationEvent.kt
@@ -2,4 +2,4 @@ package ch.epfl.sdp.replay.game_event
 
 import ch.epfl.sdp.game.data.Location
 
-class LocationEvent(timestamp: Long, val playerID: Int, val location: Location) : GameEvent(timestamp)
+class LocationEvent(timestamp: Int, val playerID: Int, val location: Location) : GameEvent(timestamp)
